requirements.txt
streamlit
transformers
torch
openai
python-dotenv
requests
beautifulsoup4
PyMuPDF
tqdm
deep-translator